Issue#245:  Extract interface RequestRunner, provide parameters in SpiceService.createRequestProcessor(), change RequestProcessor constructor
