+++ robospice-core-parent/robospice/src/main/java/com/octo/android/robospice/SpiceService.java
@@ -1,14 +1,5 @@
 package com.octo.android.robospice;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ExecutorService;
-
-import roboguice.util.temp.Ln;
 import android.app.Application;
 import android.app.Notification;
 import android.app.PendingIntent;
@@ -17,7 +8,6 @@ import android.content.Intent;
 import android.os.Binder;
 import android.os.Build;
 import android.os.IBinder;
-
 import com.octo.android.robospice.networkstate.DefaultNetworkStateChecker;
 import com.octo.android.robospice.networkstate.NetworkStateChecker;
 import com.octo.android.robospice.persistence.CacheManager;
@@ -34,6 +24,15 @@ import com.octo.android.robospice.request.listener.SpiceServiceListener;
 import com.octo.android.robospice.request.notifier.DefaultRequestListenerNotifier;
 import com.octo.android.robospice.request.notifier.RequestListenerNotifier;
 import com.octo.android.robospice.request.notifier.SpiceServiceListenerNotifier;
+import roboguice.util.temp.Ln;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.ExecutorService;
 
 /**
  * This is an abstract class used to manage the cache and provide web service
@@ -43,14 +42,18 @@ import com.octo.android.robospice.request.notifier.SpiceServiceListenerNotifier;
  * {@link #createCacheManager(Application)} to configure the
  * {@link CacheManager} used by all requests to persist their results in the
  * cache (and load them from cache if possible).
+ *
  * @author jva
  * @author mwa
  * @author sni
  */
-public abstract class SpiceService extends Service {
+public abstract class SpiceService extends Service
+{
 
     // http://stackoverflow.com/a/13359680/693752
-    /** JUNIT - this is for testing purposes only */
+    /**
+     * JUNIT - this is for testing purposes only
+     */
     private static boolean isJUnit = false;
 
     // ----------------------------------
@@ -68,7 +71,9 @@ public abstract class SpiceService extends Service {
     // ----------------------------------
     private SpiceServiceBinder mSpiceServiceBinder;
 
-    /** Responsible for processing requests. */
+    /**
+     * Responsible for processing requests.
+     */
     private RequestProcessor requestProcessor;
 
     private int currentPendingRequestCount = 0;
@@ -77,13 +82,10 @@ public abstract class SpiceService extends Service {
 
     private Notification notification;
 
-    /** Responsible for persisting data. */
+    /**
+     * Responsible for persisting data.
+     */
     private CacheManager cacheManager;
-
-    private SpiceServiceListenerNotifier spiceServiceListenerNotifier;
-    private RequestListenerNotifier progressReporter;
-
-    private final SelfStopperRequestProcessorListener requestProcessorListener = new SelfStopperRequestProcessorListener();
 
     // ----------------------------------
     // CONSTRUCTOR
@@ -92,34 +94,41 @@ public abstract class SpiceService extends Service {
     /**
      * Default constructor.
      */
-    public SpiceService() {
+    public SpiceService()
+    {
         mSpiceServiceBinder = new SpiceServiceBinder(this);
     }
 
     @Override
-    public void onCreate() {
+    public void onCreate()
+    {
         super.onCreate();
 
-        try {
+        try
+        {
             cacheManager = createCacheManager(getApplication());
-        } catch (CacheCreationException e) {
+        } catch (CacheCreationException e)
+        {
             Ln.e(e);
             stopSelf();
             return;
         }
-        if (cacheManager == null) {
+        if (cacheManager == null)
+        {
             Ln.e(new CacheCreationException("createCacheManager() can't create a null cacheManager"));
             stopSelf();
             return;
         }
 
-        progressReporter = createRequestRequestListenerNotifier();
-        spiceServiceListenerNotifier = createSpiceServiceListenerNotifier();
+        RequestListenerNotifier progressReporter = createRequestRequestListenerNotifier();
+        SpiceServiceListenerNotifier spiceServiceListenerNotifier = createSpiceServiceListenerNotifier();
+        RequestProcessorListener requestProcessorListener = createRequestProcessorListener();
 
         final ExecutorService executorService = getExecutorService();
         final NetworkStateChecker networkStateChecker = getNetworkStateChecker();
 
-        requestProcessor = createRequestProcessor(executorService, networkStateChecker);
+        requestProcessor = createRequestProcessor(executorService, networkStateChecker, requestProcessorListener,
+                cacheManager, progressReporter, spiceServiceListenerNotifier);
         requestProcessor.setFailOnCacheError(DEFAULT_FAIL_ON_CACHE_ERROR);
 
         notification = createDefaultNotification();
@@ -128,10 +137,18 @@ public abstract class SpiceService extends Service {
     }
 
     @Override
-    public int onStartCommand(final Intent intent, final int flags, final int startId) {
+    public int onStartCommand(final Intent intent, final int flags, final int startId)
+    {
         super.onStartCommand(intent, flags, startId);
         return START_NOT_STICKY;
     }
+
+
+    protected RequestProcessorListener createRequestProcessorListener()
+    {
+        return new SelfStopperRequestProcessorListener();
+    }
+
 
     /**
      * Factory method to create an entity responsible for processing requests
@@ -139,22 +156,30 @@ public abstract class SpiceService extends Service {
      * return a {@link RequestProcessor}. Override this method if you want to
      * inject a custom request processor. This feature has been implemented
      * following a request from Christopher Jenkins.
-     * @param executorService
-     *            a service executor that can be used to multi-thread request
-     *            processing.
-     * @param networkStateChecker
-     *            an entity that will check network state.
+     *
+     *
+     *
+     *
+     * @param executorService     a service executor that can be used to multi-thread request
+     *                            processing.
+     * @param networkStateChecker an entity that will check network state.
+     * @param requestProcessorListener
+     * @param progressReporter
+     * @param spiceServiceListenerNotifier
      * @return a {@link RequestProcessor} that will be used to process requests.
      */
-    protected RequestProcessor createRequestProcessor(ExecutorService executorService, NetworkStateChecker networkStateChecker) {
-        return new RequestProcessor(getApplicationContext(), cacheManager, executorService, requestProcessorListener, networkStateChecker, progressReporter, spiceServiceListenerNotifier);
+    protected RequestProcessor createRequestProcessor(ExecutorService executorService, NetworkStateChecker networkStateChecker, RequestProcessorListener requestProcessorListener, CacheManager cacheManager, RequestListenerNotifier progressReporter, SpiceServiceListenerNotifier spiceServiceListenerNotifier)
+    {
+        return RequestProcessor.createDefault(getApplicationContext(), cacheManager, executorService, requestProcessorListener, networkStateChecker, progressReporter, spiceServiceListenerNotifier);
     }
-    
+
     /**
      * For testing purposes only.
+     *
      * @return the request processor of this spice service.
      */
-    protected RequestProcessor getRequestProcessor() {
+    protected RequestProcessor getRequestProcessor()
+    {
         return requestProcessor;
     }
 
@@ -162,9 +187,11 @@ public abstract class SpiceService extends Service {
      * Method to create a Request Progress Reporter object which is responsible
      * for informing the listeners of the current state of each request. You can
      * use this method to modify the existing behavior
+     *
      * @return {@link RequestListenerNotifier}
      */
-    protected RequestListenerNotifier createRequestRequestListenerNotifier() {
+    protected RequestListenerNotifier createRequestRequestListenerNotifier()
+    {
         return new DefaultRequestListenerNotifier();
     }
 
@@ -177,10 +204,12 @@ public abstract class SpiceService extends Service {
      * related requests. In that case create a {@link NetworkStateChecker} that
      * always return true. This feature has been implemented following a request
      * from Pierre Durand.
+     *
      * @return a {@link NetworkStateChecker} that will be used to determine if
-     *         network state allows requests executions.
+     * network state allows requests executions.
      */
-    protected NetworkStateChecker getNetworkStateChecker() {
+    protected NetworkStateChecker getNetworkStateChecker()
+    {
         return new DefaultNetworkStateChecker();
     }
 
@@ -196,23 +225,30 @@ public abstract class SpiceService extends Service {
      * <p/>
      * This feature has been implemented following a request from Riccardo
      * Ciovati.
+     *
      * @return the {@link ExecutorService} to be used to execute requests .
      */
-    protected ExecutorService getExecutorService() {
+    protected ExecutorService getExecutorService()
+    {
         final int threadCount = getThreadCount();
         final int threadPriority = getThreadPriority();
-        if (threadCount <= 0) {
+        if (threadCount <= 0)
+        {
             throw new IllegalArgumentException("Thread count must be >= 1");
-        } else {
+        }
+        else
+        {
             return PriorityThreadPoolExecutor.getPriorityExecutor(threadCount, threadPriority);
         }
     }
 
     /**
      * Creates the SpiceServiceListenerNotifier.
+     *
      * @return ({@link SpiceServiceListenerNotifier)}
      */
-    protected SpiceServiceListenerNotifier createSpiceServiceListenerNotifier() {
+    protected SpiceServiceListenerNotifier createSpiceServiceListenerNotifier()
+    {
         return new SpiceServiceListenerNotifier();
     }
 
@@ -224,18 +260,25 @@ public abstract class SpiceService extends Service {
      * shown, on ICS+, a notification is shown with app icon). On Jelly Bean+,
      * the notifiation only appears when notification bar is expanded / pulled
      * down.
+     *
      * @return a notification used to tell user that the SpiceService is still
-     *         running and processing requests.
+     * running and processing requests.
      */
     @SuppressWarnings("deprecation")
-    public Notification createDefaultNotification() {
+    public Notification createDefaultNotification()
+    {
 
         Notification notification = null;
-        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN)
+        {
             notification = new Notification.Builder(this).setSmallIcon(getApplicationInfo().icon).build();
-        } else if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
+        }
+        else if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB)
+        {
             notification = new Notification.Builder(this).setSmallIcon(getApplicationInfo().icon).getNotification();
-        } else {
+        }
+        else
+        {
             notification = new Notification();
             notification.icon = getApplicationInfo().icon;
             // temporary fix https://github.com/octo-online/robospice/issues/200
@@ -245,19 +288,22 @@ public abstract class SpiceService extends Service {
             notification.when = System.currentTimeMillis();
         }
 
-        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN)
+        {
             notification.priority = Notification.PRIORITY_MIN;
         }
-        
+
         return notification;
     }
 
-    protected int getNotificationId() {
+    protected int getNotificationId()
+    {
         return DEFAULT_NOTIFICATION_ID;
     }
 
     @Override
-    public void onDestroy() {
+    public void onDestroy()
+    {
         requestProcessor.shouldStop();
         Ln.d("SpiceService instance destroyed.");
         super.onDestroy();
@@ -273,74 +319,91 @@ public abstract class SpiceService extends Service {
      * Override this method to increase the number of threads used to process
      * requests. This method will have no effect if you override
      * {@link #getExecutorService()}.
+     *
      * @return the number of threads used to process requests. Defaults to
-     *         {@link #DEFAULT_THREAD_COUNT}.
+     * {@link #DEFAULT_THREAD_COUNT}.
      */
-    public int getThreadCount() {
+    public int getThreadCount()
+    {
         return DEFAULT_THREAD_COUNT;
     }
 
     /**
      * Override this method to change the priority of threads requests. This
      * method will have no effect if you override {@link #getExecutorService()}.
+     *
      * @return the number of threads used to process requests.Defaults to
-     *         {@link #DEFAULT_THREAD_PRIORITY}.
+     * {@link #DEFAULT_THREAD_PRIORITY}.
      */
-    public int getThreadPriority() {
+    public int getThreadPriority()
+    {
         return DEFAULT_THREAD_PRIORITY;
     }
 
-    public void addRequest(final CachedSpiceRequest<?> request, final Set<RequestListener<?>> listRequestListener) {
+    public void addRequest(final CachedSpiceRequest<?> request, final Set<RequestListener<?>> listRequestListener)
+    {
         currentPendingRequestCount++;
         requestProcessor.addRequest(request, listRequestListener);
         showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification();
     }
 
-    public boolean removeDataFromCache(final Class<?> clazz, final Object cacheKey) {
+    public boolean removeDataFromCache(final Class<?> clazz, final Object cacheKey)
+    {
         return requestProcessor.removeDataFromCache(clazz, cacheKey);
     }
 
-    public void removeAllDataFromCache(final Class<?> clazz) {
+    public void removeAllDataFromCache(final Class<?> clazz)
+    {
         requestProcessor.removeAllDataFromCache(clazz);
     }
 
-    public <T> List<Object> getAllCacheKeys(final Class<T> clazz) {
+    public <T> List<Object> getAllCacheKeys(final Class<T> clazz)
+    {
         return cacheManager.getAllCacheKeys(clazz);
     }
 
-    public <T> List<T> loadAllDataFromCache(final Class<T> clazz) throws CacheLoadingException, CacheCreationException {
+    public <T> List<T> loadAllDataFromCache(final Class<T> clazz) throws CacheLoadingException, CacheCreationException
+    {
         return cacheManager.loadAllDataFromCache(clazz);
     }
 
-    public <T> T getDataFromCache(final Class<T> clazz, final Object cacheKey) throws CacheLoadingException, CacheCreationException {
+    public <T> T getDataFromCache(final Class<T> clazz, final Object cacheKey) throws CacheLoadingException, CacheCreationException
+    {
         return cacheManager.loadDataFromCache(clazz, cacheKey, DurationInMillis.ALWAYS_RETURNED);
     }
 
-    public <T> T putDataInCache(final Object cacheKey, T data) throws CacheSavingException, CacheCreationException {
+    public <T> T putDataInCache(final Object cacheKey, T data) throws CacheSavingException, CacheCreationException
+    {
         return cacheManager.saveDataToCacheAndReturnData(data, cacheKey);
     }
 
-    public boolean isDataInCache(Class<?> clazz, Object cacheKey, long cacheExpiryDuration) throws CacheCreationException {
+    public boolean isDataInCache(Class<?> clazz, Object cacheKey, long cacheExpiryDuration) throws CacheCreationException
+    {
         return cacheManager.isDataInCache(clazz, cacheKey, cacheExpiryDuration);
     }
 
-    public Date getDateOfDataInCache(Class<?> clazz, Object cacheKey) throws CacheLoadingException, CacheCreationException {
+    public Date getDateOfDataInCache(Class<?> clazz, Object cacheKey) throws CacheLoadingException, CacheCreationException
+    {
         return cacheManager.getDateOfDataInCache(clazz, cacheKey);
     }
 
-    public void removeAllDataFromCache() {
+    public void removeAllDataFromCache()
+    {
         requestProcessor.removeAllDataFromCache();
     }
 
-    public boolean isFailOnCacheError() {
+    public boolean isFailOnCacheError()
+    {
         return requestProcessor.isFailOnCacheError();
     }
 
-    public void setFailOnCacheError(final boolean failOnCacheError) {
+    public void setFailOnCacheError(final boolean failOnCacheError)
+    {
         requestProcessor.setFailOnCacheError(failOnCacheError);
     }
 
-    public void dontNotifyRequestListenersForRequest(final CachedSpiceRequest<?> request, final Collection<RequestListener<?>> listRequestListener) {
+    public void dontNotifyRequestListenersForRequest(final CachedSpiceRequest<?> request, final Collection<RequestListener<?>> listRequestListener)
+    {
         requestProcessor.dontNotifyRequestListenersForRequest(request, listRequestListener);
     }
 
@@ -349,105 +412,132 @@ public abstract class SpiceService extends Service {
     // ----------------------------------
 
     @Override
-    public IBinder onBind(final Intent intent) {
+    public IBinder onBind(final Intent intent)
+    {
         isBound = true;
         showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification();
         return mSpiceServiceBinder;
     }
 
     @Override
-    public void onRebind(Intent intent) {
+    public void onRebind(Intent intent)
+    {
         isBound = true;
         showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification();
         super.onRebind(intent);
     }
 
     @Override
-    public boolean onUnbind(final Intent intent) {
+    public boolean onUnbind(final Intent intent)
+    {
         isBound = false;
         showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification();
         stopIfNotBoundAndHasNoPendingRequests();
         return true;
     }
 
-    private final class SelfStopperRequestProcessorListener implements RequestProcessorListener {
+    private final class SelfStopperRequestProcessorListener implements RequestProcessorListener
+    {
         @Override
-        public void requestsInProgress() {
+        public void requestsInProgress()
+        {
         }
 
         @Override
-        public void allRequestComplete() {
+        public void allRequestComplete()
+        {
             currentPendingRequestCount = 0;
             stopIfNotBoundAndHasNoPendingRequests();
         }
     }
 
-    public static class SpiceServiceBinder extends Binder {
+    public static class SpiceServiceBinder extends Binder
+    {
         private final SpiceService spiceService;
 
-        public SpiceServiceBinder(final SpiceService spiceService) {
+        public SpiceServiceBinder(final SpiceService spiceService)
+        {
             this.spiceService = spiceService;
         }
 
-        public SpiceService getSpiceService() {
+        public SpiceService getSpiceService()
+        {
             return spiceService;
         }
     }
 
-    public void dumpState() {
+    public void dumpState()
+    {
         Ln.v(requestProcessor.toString());
     }
 
-    public void addSpiceServiceListener(final SpiceServiceListener spiceServiceListener) {
+    public void addSpiceServiceListener(final SpiceServiceListener spiceServiceListener)
+    {
         requestProcessor.addSpiceServiceListener(spiceServiceListener);
     }
 
-    public void removeSpiceServiceListener(final SpiceServiceListener spiceServiceListener) {
+    public void removeSpiceServiceListener(final SpiceServiceListener spiceServiceListener)
+    {
         requestProcessor.removeSpiceServiceListener(spiceServiceListener);
     }
 
-    private void stopIfNotBoundAndHasNoPendingRequests() {
+    private void stopIfNotBoundAndHasNoPendingRequests()
+    {
         Ln.v("Pending requests : " + currentPendingRequestCount);
-        if (currentPendingRequestCount == 0 && !isBound) {
+        if (currentPendingRequestCount == 0 && !isBound)
+        {
             stopSelf();
         }
     }
 
-    private void showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification() {
+    private void showNotificationIfNotBoundAndHasPendingRequestsOtherwiseHideNotification()
+    {
         // http://stackoverflow.com/a/13359680/693752
-        if (notification == null || isJUnit) {
+        if (notification == null || isJUnit)
+        {
             return;
         }
         Ln.v("Pending requests : " + currentPendingRequestCount);
-        if (isBound || currentPendingRequestCount == 0) {
+        if (isBound || currentPendingRequestCount == 0)
+        {
             Ln.v("Stop foreground");
             stopForeground(true);
-        } else {
+        }
+        else
+        {
             Ln.v("Start foreground");
             startForeground(notification);
         }
     }
 
-    private void startForeground(final Notification notification) {
-        try {
+    private void startForeground(final Notification notification)
+    {
+        try
+        {
             final Method setForegroundMethod = Service.class.getMethod("startForeground", int.class, Notification.class);
             setForegroundMethod.invoke(this, getNotificationId(), notification);
-        } catch (final SecurityException e) {
+        } catch (final SecurityException e)
+        {
             Ln.e(e, "Unable to start a service in foreground");
-        } catch (final NoSuchMethodException e) {
+        } catch (final NoSuchMethodException e)
+        {
             Ln.e(e, "Unable to start a service in foreground");
-        } catch (final IllegalArgumentException e) {
+        } catch (final IllegalArgumentException e)
+        {
             Ln.e(e, "Unable to start a service in foreground");
-        } catch (final IllegalAccessException e) {
+        } catch (final IllegalAccessException e)
+        {
             Ln.e(e, "Unable to start a service in foreground");
-        } catch (final InvocationTargetException e) {
+        } catch (final InvocationTargetException e)
+        {
             Ln.e(e, "Unable to start a service in foreground");
         }
     }
 
     // http://code.google.com/p/android/issues/detail?id=12122
     // There is a bug in ServiceTestCase : a call to setForeground will fail
-    public static final void setIsJunit(boolean b) {
+    public static final void setIsJunit(boolean b)
+    {
         isJUnit = b;
     }
 }
