+++ robospice-core-parent/robospice/src/main/java/com/octo/android/robospice/request/RequestProcessor.java
@@ -1,16 +1,6 @@
 package com.octo.android.robospice.request;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ExecutorService;
-
-import roboguice.util.temp.Ln;
 import android.content.Context;
-
 import com.octo.android.robospice.SpiceService;
 import com.octo.android.robospice.networkstate.NetworkStateChecker;
 import com.octo.android.robospice.persistence.CacheManager;
@@ -19,18 +9,24 @@ import com.octo.android.robospice.request.listener.RequestListener;
 import com.octo.android.robospice.request.listener.SpiceServiceListener;
 import com.octo.android.robospice.request.notifier.RequestListenerNotifier;
 import com.octo.android.robospice.request.notifier.SpiceServiceListenerNotifier;
+import roboguice.util.temp.Ln;
+
+import java.util.*;
+import java.util.concurrent.ExecutorService;
 
 /**
  * Delegate class of the {@link SpiceService}, easier to test than an Android
  * {@link android.app.Service}.
+ *
  * @author jva
  */
-public class RequestProcessor {
+public class RequestProcessor
+{
     // ============================================================================================
     // ATTRIBUTES
     // ============================================================================================
 
-    private final Map<CachedSpiceRequest<?>, Set<RequestListener<?>>> mapRequestToRequestListener = Collections.synchronizedMap(new LinkedHashMap<CachedSpiceRequest<?>, Set<RequestListener<?>>>());
+    private final Map<CachedSpiceRequest<?>, Set<RequestListener<?>>> mapRequestToRequestListener;
     private final RequestProgressManager requestProgressManager;
     private final RequestRunner requestRunner;
     private final CacheManager cacheManager;
@@ -43,43 +39,55 @@ public class RequestProcessor {
     /**
      * Build a request processor using a custom. This feature has been
      * implemented follwing a feature request from Riccardo Ciovati.
-     * @param context
-     *            the context on which {@link SpiceRequest} will provide their
-     *            results.
-     * @param cacheManager
-     *            the {@link CacheManager} that will be used to retrieve
-     *            requests' result and store them.
-     * @param executorService
-     *            a custom {@link ExecutorService} that will be used to execute
-     *            {@link SpiceRequest} .
-     * @param requestProcessorListener
-     *            a listener of the {@link RequestProcessor}, it will be
-     *            notified when no more requests are left, typically allowing
-     *            the {@link SpiceService} to stop itself.
+     *
+     * @param context                  the context on which {@link SpiceRequest} will provide their
+     *                                 results.
+     * @param cacheManager             the {@link CacheManager} that will be used to retrieve
+     *                                 requests' result and store them.
+     * @param executorService          a custom {@link ExecutorService} that will be used to execute
+     *                                 {@link SpiceRequest} .
+     * @param requestProcessorListener a listener of the {@link RequestProcessor}, it will be
+     *                                 notified when no more requests are left, typically allowing
+     *                                 the {@link SpiceService} to stop itself.
      */
-    public RequestProcessor(final Context context, final CacheManager cacheManager, final ExecutorService executorService, final RequestProcessorListener requestProcessorListener,
-        final NetworkStateChecker networkStateChecker, final RequestListenerNotifier requestListenerNotifier, final SpiceServiceListenerNotifier spiceServiceListenerNotifier) {
+    public static RequestProcessor createDefault(final Context context, final CacheManager cacheManager, final ExecutorService executorService, final RequestProcessorListener requestProcessorListener, final NetworkStateChecker networkStateChecker, final RequestListenerNotifier requestListenerNotifier, final SpiceServiceListenerNotifier spiceServiceListenerNotifier)
+    {
+        Map<CachedSpiceRequest<?>, Set<RequestListener<?>>> mapRequestToRequestListener = Collections.synchronizedMap(new LinkedHashMap<CachedSpiceRequest<?>, Set<RequestListener<?>>>());
+        RequestProgressManager requestProgressManager = new RequestProgressManager(requestProcessorListener, mapRequestToRequestListener, requestListenerNotifier, spiceServiceListenerNotifier);
+        RequestRunner requestRunner = new DefaultRequestRunner(context, cacheManager, executorService, requestProgressManager, networkStateChecker);
+        return new RequestProcessor(mapRequestToRequestListener, requestProgressManager, requestRunner, cacheManager);
+    }
 
+    public RequestProcessor(Map<CachedSpiceRequest<?>, Set<RequestListener<?>>> mapRequestToRequestListener,
+                            RequestProgressManager requestProgressManager, RequestRunner requestRunner, CacheManager cacheManager)
+    {
+        this.mapRequestToRequestListener = mapRequestToRequestListener;
+        this.requestProgressManager = requestProgressManager;
+        this.requestRunner = requestRunner;
         this.cacheManager = cacheManager;
-        this.requestProgressManager = new RequestProgressManager(requestProcessorListener, mapRequestToRequestListener, requestListenerNotifier, spiceServiceListenerNotifier);
-        this.requestRunner = new RequestRunner(context, cacheManager, executorService, requestProgressManager, networkStateChecker);
     }
 
     // ============================================================================================
     // PUBLIC
     // ============================================================================================
-    public void addRequest(final CachedSpiceRequest<?> request, final Set<RequestListener<?>> listRequestListener) {
-        if (isStopped) {
+    public void addRequest(final CachedSpiceRequest<?> request, final Set<RequestListener<?>> listRequestListener)
+    {
+        if (isStopped)
+        {
             Ln.d("Dropping request : " + request + " as processor is stopped.");
             return;
         }
 
         Ln.d("Adding request to queue " + hashCode() + ": " + request + " size is " + mapRequestToRequestListener.size());
 
-        if (request.isCancelled()) {
-            synchronized (mapRequestToRequestListener) {
-                for (final CachedSpiceRequest<?> cachedSpiceRequest : mapRequestToRequestListener.keySet()) {
-                    if (cachedSpiceRequest.equals(request)) {
+        if (request.isCancelled())
+        {
+            synchronized (mapRequestToRequestListener)
+            {
+                for (final CachedSpiceRequest<?> cachedSpiceRequest : mapRequestToRequestListener.keySet())
+                {
+                    if (cachedSpiceRequest.equals(request))
+                    {
                         cachedSpiceRequest.cancel();
                         requestProgressManager.notifyListenersOfRequestCancellation(request);
                         return;
@@ -91,30 +99,40 @@ public class RequestProcessor {
         boolean aggregated = false;
         Set<RequestListener<?>> listRequestListenerForThisRequest = mapRequestToRequestListener.get(request);
 
-        if (listRequestListenerForThisRequest == null) {
-            if (request.isProcessable()) {
+        if (listRequestListenerForThisRequest == null)
+        {
+            if (request.isProcessable())
+            {
                 Ln.d("Adding entry for type %s and cacheKey %s.", request.getResultType(), request.getRequestCacheKey());
                 listRequestListenerForThisRequest = Collections.synchronizedSet(new HashSet<RequestListener<?>>());
                 this.mapRequestToRequestListener.put(request, listRequestListenerForThisRequest);
             }
-        } else {
+        }
+        else
+        {
             Ln.d("Request for type %s and cacheKey %s already exists.", request.getResultType(), request.getRequestCacheKey());
             aggregated = true;
         }
 
-        if (listRequestListener != null && listRequestListenerForThisRequest != null) {
+        if (listRequestListener != null && listRequestListenerForThisRequest != null)
+        {
             listRequestListenerForThisRequest.addAll(listRequestListener);
         }
 
-        if (aggregated) {
+        if (aggregated)
+        {
             requestProgressManager.notifyListenersOfRequestAggregated(request, listRequestListener);
             return;
         }
 
-        if (request.isProcessable()) {
+        if (request.isProcessable())
+        {
             requestProgressManager.notifyListenersOfRequestAdded(request, listRequestListener);
-        } else {
-            if (listRequestListenerForThisRequest == null) {
+        }
+        else
+        {
+            if (listRequestListenerForThisRequest == null)
+            {
                 requestProgressManager.notifyListenersOfRequestNotFound(request, listRequestListener);
             }
             requestProgressManager.notifyOfRequestProcessed(request, listRequestListener);
@@ -123,21 +141,26 @@ public class RequestProcessor {
             return;
         }
 
-        final RequestCancellationListener requestCancellationListener = new RequestCancellationListener() {
+        final RequestCancellationListener requestCancellationListener = new RequestCancellationListener()
+        {
 
             @Override
-            public void onRequestCancelled() {
+            public void onRequestCancelled()
+            {
                 requestProgressManager.notifyListenersOfRequestCancellation(request);
                 mapRequestToRequestListener.remove(request);
             }
         };
         request.setRequestCancellationListener(requestCancellationListener);
 
-        if (request.isCancelled()) {
+        if (request.isCancelled())
+        {
             requestProgressManager.notifyListenersOfRequestCancellation(request);
             mapRequestToRequestListener.remove(request);
             return;
-        } else {
+        }
+        else
+        {
             requestRunner.executeRequest(request);
         }
     }
@@ -146,38 +169,44 @@ public class RequestProcessor {
      * Disable request listeners notifications for a specific request.<br/>
      * All listeners associated to this request won't be called when request
      * will finish.<br/>
-     * @param request
-     *            Request on which you want to disable listeners
-     * @param listRequestListener
-     *            the collection of listeners associated to request not to be
-     *            notified
+     *
+     * @param request             Request on which you want to disable listeners
+     * @param listRequestListener the collection of listeners associated to request not to be
+     *                            notified
      */
-    public void dontNotifyRequestListenersForRequest(final CachedSpiceRequest<?> request, final Collection<RequestListener<?>> listRequestListener) {
+    public void dontNotifyRequestListenersForRequest(final CachedSpiceRequest<?> request, final Collection<RequestListener<?>> listRequestListener)
+    {
         requestProgressManager.dontNotifyRequestListenersForRequest(request, listRequestListener);
     }
 
-    public boolean removeDataFromCache(final Class<?> clazz, final Object cacheKey) {
+    public boolean removeDataFromCache(final Class<?> clazz, final Object cacheKey)
+    {
         return cacheManager.removeDataFromCache(clazz, cacheKey);
     }
 
-    public void removeAllDataFromCache(final Class<?> clazz) {
+    public void removeAllDataFromCache(final Class<?> clazz)
+    {
         cacheManager.removeAllDataFromCache(clazz);
     }
 
-    public void removeAllDataFromCache() {
+    public void removeAllDataFromCache()
+    {
         cacheManager.removeAllDataFromCache();
     }
 
-    public boolean isFailOnCacheError() {
+    public boolean isFailOnCacheError()
+    {
         return requestRunner.isFailOnCacheError();
     }
 
-    public void setFailOnCacheError(final boolean failOnCacheError) {
+    public void setFailOnCacheError(final boolean failOnCacheError)
+    {
         requestRunner.setFailOnCacheError(failOnCacheError);
     }
 
     @Override
-    public String toString() {
+    public String toString()
+    {
         final StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append('[');
         stringBuilder.append(getClass().getName());
@@ -187,14 +216,18 @@ public class RequestProcessor {
         stringBuilder.append(mapRequestToRequestListener.keySet().size());
 
         stringBuilder.append(", listeners per requests = [");
-        for (final Map.Entry<CachedSpiceRequest<?>, Set<RequestListener<?>>> entry : mapRequestToRequestListener.entrySet()) {
+        for (final Map.Entry<CachedSpiceRequest<?>, Set<RequestListener<?>>> entry : mapRequestToRequestListener.entrySet())
+        {
             stringBuilder.append(entry.getKey().getClass().getName());
             stringBuilder.append(":");
             stringBuilder.append(entry.getKey());
             stringBuilder.append(" --> ");
-            if (entry.getValue() == null) {
+            if (entry.getValue() == null)
+            {
                 stringBuilder.append(entry.getValue());
-            } else {
+            }
+            else
+            {
                 stringBuilder.append(entry.getValue().size());
             }
         }
@@ -204,20 +237,24 @@ public class RequestProcessor {
         return stringBuilder.toString();
     }
 
-    public void addSpiceServiceListener(final SpiceServiceListener spiceServiceListener) {
+    public void addSpiceServiceListener(final SpiceServiceListener spiceServiceListener)
+    {
         requestProgressManager.addSpiceServiceListener(spiceServiceListener);
     }
 
-    public void removeSpiceServiceListener(final SpiceServiceListener spiceServiceListener) {
+    public void removeSpiceServiceListener(final SpiceServiceListener spiceServiceListener)
+    {
         requestProgressManager.removeSpiceServiceListener(spiceServiceListener);
     }
 
-    public void shouldStop() {
+    public void shouldStop()
+    {
         isStopped = true;
         requestRunner.shouldStop();
     }
 
-    public boolean isStopped() {
+    public boolean isStopped()
+    {
         return isStopped;
     }
 }
